+++ app/registration/confirmCtrl.js
@@ -14,7 +14,7 @@
             function confirmUser () {
                 commonService.confirmUser(vm.userDetails)
                     .then(function (response) {
-                        vm.message.value = 'Thank you for confirming your account. You may log in.';
+                        vm.message.value = 'Thank you for confirming your account. You may now log in.';
                         vm.message.success = true;
                     },function (error) {
                         vm.message.value = error.data.error;